requirements.txt
Flask
mysql-connector-python
python-dotenv
google-genai
pillow
markdown